+++ tests/testthat/test-interact.R
@@ -338,6 +338,48 @@ test_that("empty selection tidy method works", {
   expect_identical(tidy(rec, number = 1), expect)
 })
 
+test_that("keep_original_cols works", {
+  new_names <- c("x1_x_x2")
+
+  rec <- recipe(~ x1 + x2, data = dat_tr) %>%
+    step_interact(~ x1:x2, keep_original_cols = FALSE)
+
+  rec <- prep(rec)
+  res <- bake(rec, new_data = NULL)
+
+  expect_equal(
+    colnames(res),
+    new_names
+  )
+
+  rec <- recipe(~ x1 + x2, data = dat_tr) %>%
+    step_interact(~ x1:x2, keep_original_cols = TRUE)
+
+  rec <- prep(rec)
+  res <- bake(rec, new_data = NULL)
+
+  expect_equal(
+    colnames(res),
+    c("x1", "x2", new_names)
+  )
+})
+
+test_that("keep_original_cols - can prep recipes with it missing", {
+  rec <- recipe(~ x1 + x2, data = dat_tr) %>%
+    step_interact(~ x1:x2, keep_original_cols = FALSE)
+
+  rec$steps[[1]]$keep_original_cols <- NULL
+
+  expect_snapshot(
+    rec <- prep(rec)
+  )
+
+  expect_error(
+    bake(rec, new_data = dat_tr),
+    NA
+  )
+})
+
 test_that("printing", {
   rec <- recipe(y ~ ., data = dat_tr) %>%
     step_interact(~ x1:x2)